ml-server/requirements.txt
fastapi
uvicorn
ultralytics
tensorflow
pillow
numpy
torch
torchvision
timm
